+++ service/controller/resource/clusterconfigmap/resource.go
@@ -1,6 +1,7 @@
 package clusterconfigmap
 
 import (
+	"fmt"
 	"strings"
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
@@ -47,8 +48,7 @@ type Resource struct {
 // New creates a new configured config map state getter resource managing
 // cluster config maps.
 //
-//     https://pkg.go.dev/github.com/giantswarm/operatorkit/v4/pkg/resource/k8s/secretresource#StateGetter
-//
+//	https://pkg.go.dev/github.com/giantswarm/operatorkit/v4/pkg/resource/k8s/secretresource#StateGetter
 func New(config Config) (*Resource, error) {
 	if config.K8sClient == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.K8sClient must not be empty", config)
@@ -92,4 +92,8 @@ func New(config Config) (*Resource, error) {
 
 func (r *Resource) Name() string {
 	return Name
+}
+
+func remoteWriteUrl(baseDomain, clusterID string) string {
+	return fmt.Sprintf("https://prometheus.g8s.%s/%s/api/v1/write", baseDomain, clusterID)
 }